backend/requirements.txt
fastapi
uvicorn
selenium
python-dotenv
supabase
pinecone-client
langchain
langchain-huggingface
langchain-pinecone
langchain-google-genai 
langchain-core
beautifulsoup4
requests
pydantic-settings
sentence-transformers 
google-generativeai
huggingface-hub
transformers   
langchain-core
tdqm
langchain-community
langgraph
PyMuPDF
python-multipart